requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
httpx==0.27.0
python-dateutil==2.9.0
Jinja2==3.1.3
